openadr_backend/pyproject.toml
[tool.poetry]
name = "openadr-backend"
version = "0.1.0"
description = "OpenADR VTN Admin Backend"
authors = ["Siddharth Shah"]
readme = "README.md"
packages = [{ include = "app" }] 

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
python-dotenv = "^1.0.0"
pydantic = "^2.6.0"
pydantic-settings = "^2.2.1"
sqlmodel = "^0.0.24"
python-jose = {version = "^3.3", extras = ["cryptography"]}
passlib = {version = "^1.7", extras = ["bcrypt"]}
boto3 = "^1.34"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.6"
aiosqlite = "^0.20.0"

[tool.poetry.scripts]
start = "uvicorn app.main:app --host 0.0.0.0 --port 8000 --reload"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
